패키지 구조 변경 및 회원가입 로직 개발
