--- src/main/java/flab/gotable/mapper/MemberMapper.java
@@ -1,9 +1,0 @@
-package flab.gotable.mapper;
-
-import flab.gotable.member.dto.request.MemberSignUpRequestDto;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface MemberMapper {
-    public void saveMember(MemberSignUpRequestDto memberSignUpRequestDto);
-}
