+++ src/main/resources/mapper/MemberMapper.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="flab.gotable.mapper.MemberMapper">
-    <insert id="saveMember" parameterType="flab.gotable.member.dto.request.MemberSignUpRequestDto">
+<mapper namespace="flab.gotable.Repository.MemberRepository">
+    <insert id="saveMember" parameterType="flab.gotable.dto.request.MemberSignUpRequestDto">
         INSERT INTO member
         (SEQ, NAME, ID, PASSWORD, PHONE)
         VALUES(NULL, #{name}, #{id}, #{password}, #{phone})
